+++ providers.go
@@ -101,21 +101,33 @@ type SlogLoggerf struct {
 
 // Debugf implements Debugf method for LogfProvider interface
 func (s *SlogLoggerf) Debugf(ctx context.Context, format string, v ...any) {
+	if !s.Logger.Enabled(ctx, slog.LevelDebug) {
+		return
+	}
 	s.Logger.DebugContext(ctx, fmt.Sprintf(format, v...))
 }
 
 // Infof implements Infof method for LogfProvider interface
 func (s SlogLoggerf) Infof(ctx context.Context, format string, v ...any) {
+	if !s.Logger.Enabled(ctx, slog.LevelInfo) {
+		return
+	}
 	s.Logger.InfoContext(ctx, fmt.Sprintf(format, v...))
 }
 
 // Warnf implements Warnf method for LogfProvider interface
 func (s SlogLoggerf) Warnf(ctx context.Context, format string, v ...any) {
+	if !s.Logger.Enabled(ctx, slog.LevelWarn) {
+		return
+	}
 	s.Logger.WarnContext(ctx, fmt.Sprintf(format, v...))
 }
 
 // Errorf implements Errorf method for LogfProvider interface
 func (s SlogLoggerf) Errorf(ctx context.Context, format string, v ...any) {
+	if !s.Logger.Enabled(ctx, slog.LevelError) {
+		return
+	}
 	s.Logger.ErrorContext(ctx, fmt.Sprintf(format, v...))
 }